Minor refactoring (mainly to adding courses)
- move logic related to add courses out of AppStoreActions
- change Schedules.AddCourse to use existing course if exists, and if not assign color
- fix addSchedule to change current schedule number to newest created schedule
- remove color argument from AppStoreActions.AddCourse because it is always undefined
