+++ src/components/RightPane/SectionTable/SectionTableButtons.tsx
@@ -86,7 +86,7 @@ export const ScheduleAddCell = withStyles(styles)((props: ScheduleAddCellProps) 
                     action: analyticsEnum.classSearch.actions.ADD_SPECIFIC,
                 });
             }
-            section.color = addCourse(section, courseDetails, term, scheduleIndex);
+            addCourse(section, courseDetails, term, scheduleIndex);
         }
     };
